mentorcontroller에 add폼은 뜨는데 add함수에 안들어감.

+++ src/main/java/pnu/ibe/justice/mentoring/repos/MentorRepository.java
@@ -5,7 +5,7 @@ import pnu.ibe.justice.mentoring.domain.Mentor;
 import pnu.ibe.justice.mentoring.domain.User;
 
 
-public interface MentorRepository extends JpaRepository<Mentor, Long> {
+public interface MentorRepository extends JpaRepository<Mentor, Integer> {
 
     Mentor findFirstByUsers(User user);
